requirements.txt
click~=8.1.3
halo~=0.0.31
pyenchant~=3.2.2
python-dotenv~=1.0.0
requests==2.29.0
Send2Trash==1.8.2
psutil==5.9.5
beaupy~=3.5.3
